Spotlight Gulp Initial Version
This is the initial commit for the new internal spotlight gulp tool that
we will use to generate static templates for PSD > HTML projects.
Further documentation will follow but the tools of importance are:
nunjucks for templating, node sass for sass compilation and rollupjs for
js bundling. A small change will follow to handle image minification.